Refactoring Code

+++ app/src/main/java/com/ph03nix_x/capacityinfo/TimeSpanInterface.kt
@@ -2,9 +2,9 @@ package com.ph03nix_x.capacityinfo
 
 interface TimeSpanInterface {
 
-    fun toHours(seconds: Double) = if(seconds >= 3600) ((seconds / 3600) % 60).toInt() else 0
+    fun toHours(seconds: Int) = if(seconds >= 3600) ((seconds / 3600) % 60) else 0
 
-    fun toMinutes(seconds: Double) = if(seconds >= 60) (((seconds / 60) - (toHours(seconds) * 60)) % 60).toInt() else 0
+    fun toMinutes(seconds: Int) = if(seconds >= 60) (((seconds / 60) - (toHours(seconds) * 60)) % 60) else 0
 
-    fun toSeconds(seconds: Double) = ((seconds - (toMinutes(seconds) * 60) - (toHours(seconds) * 3600)) % 60).toInt()
+    fun toSeconds(seconds: Int) = ((seconds - (toMinutes(seconds) * 60) - (toHours(seconds) * 3600)) % 60)
 }
